+++ app/src/handy_search.h
@@ -15,7 +15,7 @@ public:
 signals:
   void startLoading(int from=1);
   void query(const QString &sentence);
-  void progress(float progress);
+  void progress(double progress);
   void loadFinished();
 protected:
   void resizeEvent(QResizeEvent *event);
